% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor.R
\name{bayes_factor}
\alias{bayes_factor}
\title{Compute Bayes factors for TWAS Z-scores}
\usage{
bayes_factor(
  zscores,
  idx_set,
  wcor,
  prior_chisq = 40,
  prb = 0.001,
  use_log = T
)
}
\arguments{
\item{zscores}{numeric, vector of Z-scores}

\item{idx_set}{character, ids of transcripts}

\item{wcor}{numeric, correlation matrix from estimate_cor}

\item{prior_chisq}{numeric, prior for chi-square mean}

\item{prb}{numeric, prior for causal configuration}

\item{use_log}{logical, use log of bayes factors}
}
\value{
bayes factor vector
}
\description{
The function runs a modified version of FOCUS's Bayes factor
calculation
}
