% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burdenTest.R
\name{burdenTest_all}
\alias{burdenTest_all}
\title{Run burden test on all transcripts in an isotwas_model}
\usage{
burdenTest_all(
  model,
  ld,
  gene,
  sumStats,
  chr = NULL,
  pos = NULL,
  a1,
  a2,
  snpName = "SNP",
  Z = NULL,
  beta = NULL,
  se = NULL,
  R2cutoff = 0.01,
  alpha = 2.5e-06,
  nperms = 1000,
  usePos = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{model}{An isotwas_model or isotwas_result object}

\item{ld}{matrix, ld reference matrix with SNP names as row/colnames}

\item{gene}{character, gene name}

\item{sumStats}{data frame, GWAS summary statistics}

\item{chr}{character, colnames in sumStats that keeps the chromosome}

\item{pos}{character, colnames in sumStats that keeps the position}

\item{a1}{character, colnames in sumStats that keeps the ALT allele}

\item{a2}{character, colnames in sumStats that keeps the REF allele}

\item{snpName}{character, colnames in sumStats that keeps the SNP id}

\item{Z}{character, colnames in sumStats that keeps the Z score}

\item{beta}{character, colnames in sumStats that keeps the effect size}

\item{se}{character, colnames in sumStats that keeps the standard error}

\item{R2cutoff}{numeric, predictive R2 cutoff}

\item{alpha}{numeric, P-value threshold for permutation testing}

\item{nperms}{numeric, number of permutations}

\item{usePos}{logical, use SNP positions vs. SNP ids}

\item{verbose}{logical, print progress}
}
\value{
data.frame with burden test results for all transcripts
}
\description{
Convenience function to run burdenTest on all transcripts in an
isotwas_model or isotwas_result object.
}
