% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{$.isotwas_model}
\alias{$.isotwas_model}
\title{Extract a transcript model from isotwas_model}
\usage{
\method{$}{isotwas_model}(x, name)
}
\arguments{
\item{x}{An isotwas_model object}

\item{name}{Transcript name to extract}
}
\description{
Extract a transcript model from isotwas_model
}
