% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_blup.R
\name{compute_blup}
\alias{compute_blup}
\title{Univariate BLUP on all columns}
\usage{
compute_blup(
  X,
  Y,
  Omega,
  scale = F,
  nfolds = 5,
  verbose,
  par = F,
  n.cores = NULL,
  tx_names = NULL,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{Omega}{matrix, precision matrix of Y}

\item{scale}{logical, T/F to scale Y by Omega}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{tx_names}{vector, character vector of tx names in order of columns of Y}

\item{seed}{int, random seed}
}
\value{
data frame of elastic net, lasso, and LMM based predictions
}
\description{
The function trains unvariate rrBLUP models individually for all
isoform transcripts
}
