% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_curds_whey.R
\name{compute_curds_whey}
\alias{compute_curds_whey}
\title{Breiman and Friedman's curds and whey multivariate regression with CV}
\usage{
compute_curds_whey(
  X,
  Y,
  family = "gaussian",
  alpha = 0.5,
  nfolds = 5,
  verbose,
  par = F,
  n.cores = NULL,
  tx_names = NULL,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{family}{character, glmnet glm family}

\item{alpha}{numeric, elastic net mixing parameter}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{tx_names}{vector, character vector of tx names - order of columns of Y}

\item{seed}{int, random seed}

\item{scale}{logical, T/F to scale Y by Omega}
}
\value{
data frame of elastic net, lasso, and LMM based predictions
}
\description{
The function trains the curds and whey multivariate regression with
cross-validated results
}
