% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_finemap_regress.R
\name{compute_finemap_regress}
\alias{compute_finemap_regress}
\title{Fine-mapping and clustering SEs with SUR}
\usage{
compute_finemap_regress(
  X,
  Y,
  Y.rep,
  R,
  id,
  nfolds,
  verbose = F,
  tx_names = NULL,
  coverage = 0.9,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G mean isoform expression across columns}

\item{Y.rep}{matrix, matrix of G isoform expression with replicates}

\item{R}{int, number of replicates}

\item{id}{vector, vector of sample ids showing rep to id}

\item{verbose}{logical}

\item{tx_names}{vector, character vector of tx names - order of columns of Y}

\item{seed}{int, random seed}
}
\value{
isoTWAS model list
}
\description{
The function feature selects with SuSiE and runs SUR with cluster-robust
SEs
}
