% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fixed.R
\name{compute_fixed}
\alias{compute_fixed}
\title{Multivariate regression with covariance estimate}
\usage{
compute_fixed(X, Y, lam2, Omega, tol.in, maxit.in, silent)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{lam2}{numeric, lambda penalty for LASSO}

\item{Omega}{matrix, precision matrix of Y}

\item{tol.in}{numeric, tolerance for objective difference}

\item{maxit.in}{int, maximum number of iteractions}

\item{silent}{logical}

\item{weights}{matrix, weights for each predictor in X}
}
\value{
list of MRCE estimates
}
\description{
The function trains multivariate lasso with a given covariance estimate
}
