% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isotwas.R
\name{compute_isotwas}
\alias{compute_isotwas}
\title{Compute isoTWAS model for a set of isoforms/transcripts}
\usage{
compute_isotwas(
  X,
  Y,
  gene_exp = NULL,
  Y.rep,
  R = NULL,
  id,
  omega_est = "replicates",
  omega_nlambda = 10,
  method = c("mrce_lasso", "multi_enet", "joinet", "spls", "sgl", "mtlasso", "stacking",
    "graph_reg", "univariate"),
  similarity_matrix = NULL,
  predict_nlambda = 20,
  family = "gaussian",
  scale = FALSE,
  alpha = 0.5,
  nfolds = 5,
  verbose = TRUE,
  par = FALSE,
  n.cores = NULL,
  tx_names = NULL,
  seed = NULL,
  run_all = TRUE,
  return_all = FALSE,
  tol.in = 0.001,
  maxit.in = 1000
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{gene_exp}{vector, vector of total gene expression}

\item{Y.rep}{matrix, matrix of G isoform expression with replicates}

\item{R}{int, number of replicates}

\item{id}{vector, vector of sample ids showing rep to id}

\item{omega_est}{character, 'replicates' or 'mean' to use Y.rep or Y}

\item{omega_nlambda}{int, number of omegas to generate}

\item{method}{character, vector of methods to use. Available methods:
\itemize{
  \item "mrce_lasso": MRCE with lasso penalty
  \item "multi_enet": Multivariate elastic net
  \item "joinet": Joinet stacked elastic net
  \item "spls": Sparse partial least squares
  \item "sgl": Sparse group lasso
  \item "mtlasso": Multi-task lasso (L21 regularization)
  \item "stacking": Super learner stacking of multiple methods
  \item "graph_reg": Graph-regularized regression (requires similarity_matrix)
  \item "univariate": Best of univariate elastic net, BLUP, SuSiE
}}

\item{similarity_matrix}{matrix, optional similarity matrix for graph_reg method.
Can be created using \code{\link{similarity_from_gtf}} or
\code{\link{create_similarity_from_exons}}. If NULL, graph_reg is skipped.}

\item{predict_nlambda}{int, number of lambdas in MRCE}

\item{family}{character, glmnet family}

\item{scale}{logical, T/F to scale Y by Omega}

\item{alpha}{numeric, elastic net mixing parameter}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical, print progress messages}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{tx_names}{vector, character vector of tx names - order of columns of Y}

\item{seed}{int, random seed}

\item{run_all}{logical, run all methods}

\item{return_all}{logical, return R2 for all models?}

\item{tol.in}{numeric, tolerance for objective difference}

\item{maxit.in}{int, maximum number of iterations}
}
\value{
optimal isoTWAS model
}
\description{
The function runs multivariate regression models for a set of transcripts
and outputs the best model based on cross-validated R-squared.
}
