% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mrce.R
\name{compute_mrce}
\alias{compute_mrce}
\title{Multivariate regression with covariance estimate}
\usage{
compute_mrce(
  X,
  Y,
  lambda = NULL,
  nlambda = 50,
  Omega,
  nfolds = 5,
  tol.in,
  maxit.in,
  verbose,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{lambda}{vector, lambda penalty vector for LASSO to tune on}

\item{Omega}{matrix, precision matrix of Y}

\item{nfolds}{int, number of CV folds}

\item{tol.in}{numeric, tolerance for objective difference}

\item{maxit.in}{int, maximum number of iteractions}

\item{verbose}{logical}

\item{seed}{int, random seed}
}
\value{
CV MRCE fit
}
\description{
The function trains multivariate lasso with a given covariance estimate
}
