% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_omega.R
\name{compute_omega}
\alias{compute_omega}
\title{Compute precision matrix between isoforms}
\usage{
compute_omega(
  Y,
  Y.rep,
  R,
  id,
  method = c("replicates", "mean"),
  nlambda = 15,
  verbose = F
)
}
\arguments{
\item{Y}{matrix, matrix of G mean isoform expression across columns}

\item{Y.rep}{matrix, matrix of G isoform expression with replicates}

\item{R}{int, number of replicates}

\item{id}{vector, vector of sample ids showing rep to id}

\item{method}{character, use Y.rep or Y}

\item{verbose}{logical}
}
\value{
covariance and/or precision matrix list in huge form
}
\description{
The function computes the covariance and precision matrix between
columns of the isoform matrix
}
