% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spring.R
\name{compute_spring}
\alias{compute_spring}
\title{Multivariate regression with covariance estimate using elastic net}
\usage{
compute_spring(
  X,
  Y,
  Omega,
  intercept = F,
  normalize = T,
  nfolds = 5,
  tol.in,
  maxit.in,
  verbose = T,
  par = F,
  n.cores,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{Omega}{matrix, precision matrix of Y}

\item{intercept}{logical, include intercept}

\item{normalize}{logical, normalize X and Y}

\item{nfolds}{int, number of CV folds}

\item{tol.in}{numeric, tolerance for objective difference}

\item{maxit.in}{int, maximum number of iterations}

\item{verbose}{logical}

\item{par}{logical, to parallelize}

\item{seed}{int, random seed}

\item{ncores}{int, number of cores}
}
\value{
CV MRCE fit
}
\description{
The function trains multivariate elastic-net with a given
covariance estimate
}
