% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_test_focus.R
\name{conditional_test_focus}
\alias{conditional_test_focus}
\title{Conditional gene-level test}
\usage{
conditional_test_focus(w_gene, w_iso, ld, z, gene_name, tx_name)
}
\arguments{
\item{w_gene}{vector, weights for gene}

\item{w_iso}{matrix, weights for significant isoforms}

\item{ld}{matrix, LD matrix}

\item{z}{vector, GWAS Z-scores}

\item{gene_name}{character, gene name}

\item{tx_name}{vector, character vector of tx names in order of columns of Y}
}
\value{
data frame of conditional Z-scores for all combinations of tx
}
\description{
The function tests the gene association with trait conditional
on isoforms
}
