% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{convert_legacy_model}
\alias{convert_legacy_model}
\title{Convert legacy model list to isotwas_model}
\usage{
convert_legacy_model(model_list, method, n_samples, n_snps)
}
\arguments{
\item{model_list}{List in legacy format (list of lists with Transcript, Model, R2, P, Pred)}

\item{method}{Character string identifying the method}

\item{n_samples}{Number of samples}

\item{n_snps}{Total number of SNPs}
}
\value{
An isotwas_model object
}
\description{
Converts the old list-of-lists format to the new isotwas_model class
}
