% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{create_isotwas_model}
\alias{create_isotwas_model}
\title{isoTWAS Model Classes and Methods}
\usage{
create_isotwas_model(method, transcripts, n_samples, n_snps)
}
\arguments{
\item{method}{character, name of the prediction method}

\item{transcripts}{named list of transcript model results}

\item{n_samples}{int, number of samples used for training}

\item{n_snps}{int, total number of SNPs in input}
}
\value{
An isotwas_model object
}
\description{
S3 classes for representing isoTWAS prediction model outputs with
intuitive structure and accessor methods.
Create an isotwas_model object
}
