% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{create_isotwas_result}
\alias{create_isotwas_result}
\title{Create an isotwas_result object}
\usage{
create_isotwas_result(
  best_models,
  method_results = NULL,
  comparison_df = NULL,
  tx2gene = NULL
)
}
\arguments{
\item{best_models}{Named list of best models per transcript}

\item{method_results}{List of isotwas_model objects (one per method)}

\item{comparison_df}{Data frame comparing methods}

\item{tx2gene}{Transcript to gene coefficients}
}
\value{
An isotwas_result object
}
\description{
Create an isotwas_result object
}
