% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_stacking.R
\name{.combine_base_weights}
\alias{.combine_base_weights}
\title{Combine base model weights using stacking weights}
\usage{
.combine_base_weights(
  final_base_models,
  stack_weights,
  base_methods,
  tx_names,
  snp_names
)
}
\description{
Combine base model weights using stacking weights
}
\keyword{internal}
