% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_from_gtf.R
\name{.compute_exon_similarity}
\alias{.compute_exon_similarity}
\title{Compute similarity between two sets of exons}
\usage{
.compute_exon_similarity(exons_i, exons_j, method)
}
\description{
Compute similarity between two sets of exons
}
\keyword{internal}
