% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_sgl.R
\name{.compute_lambda_max_sgl}
\alias{.compute_lambda_max_sgl}
\title{Compute lambda_max for Sparse Group LASSO}
\usage{
.compute_lambda_max_sgl(X, Y, alpha)
}
\description{
Compute lambda_max for Sparse Group LASSO
}
\keyword{internal}
