% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_from_gtf.R
\name{.find_gene_in_gtf}
\alias{.find_gene_in_gtf}
\title{Find gene in GTF with flexible matching}
\usage{
.find_gene_in_gtf(gtf, target_gene)
}
\description{
Find gene in GTF with flexible matching
}
\keyword{internal}
