% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_stacking.R
\name{.fit_base_method_coef}
\alias{.fit_base_method_coef}
\title{Get coefficients from base method using pre-computed hyperparameters}
\usage{
.fit_base_method_coef(method, X, Y, params, seed)
}
\description{
Get coefficients from base method using pre-computed hyperparameters
}
\keyword{internal}
