% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_stacking.R
\name{.fit_base_method_fast}
\alias{.fit_base_method_fast}
\title{Fast base method fitting using pre-computed hyperparameters}
\usage{
.fit_base_method_fast(method, X_train, Y_train, X_test, params, seed)
}
\description{
Fast base method fitting using pre-computed hyperparameters
}
\keyword{internal}
