% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_mtlasso.R
\name{.fit_mtl_l21}
\alias{.fit_mtl_l21}
\title{Fit L21-penalized multi-task regression via block coordinate descent}
\usage{
.fit_mtl_l21(
  X,
  Y,
  lambda,
  XtX_diag = NULL,
  XtY = NULL,
  W_init = NULL,
  max_iter = 500,
  tol = 1e-04
)
}
\description{
Fit L21-penalized multi-task regression via block coordinate descent
}
\keyword{internal}
