% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_sgl.R
\name{.fit_sgl}
\alias{.fit_sgl}
\title{Fit Sparse Group LASSO using coordinate descent}
\usage{
.fit_sgl(X, Y, lambda, alpha, B_init = NULL, max_iter = 500, tol = 1e-04)
}
\description{
Fit Sparse Group LASSO using coordinate descent
}
\keyword{internal}
