% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_mtlasso.R
\name{.mtlasso_rmtl}
\alias{.mtlasso_rmtl}
\title{RMTL backend for Multi-Task LASSO}
\usage{
.mtlasso_rmtl(
  X,
  Y,
  regularization = "L21",
  Lam1_seq = NULL,
  Lam2 = 0,
  nfolds = 5,
  verbose = FALSE,
  seed = 123
)
}
\description{
Uses the RMTL package. Slower but supports Trace and Lasso regularization.
}
\keyword{internal}
