% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_stacking.R
\name{.nnls_weights_fast}
\alias{.nnls_weights_fast}
\title{Fast non-negative least squares for stacking weights}
\usage{
.nnls_weights_fast(Z, y)
}
\description{
Fast non-negative least squares for stacking weights
}
\keyword{internal}
