% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_stacking.R
\name{.precompute_hyperparams}
\alias{.precompute_hyperparams}
\title{Precompute optimal hyperparameters for each method}
\usage{
.precompute_hyperparams(X, Y, base_methods, seed)
}
\description{
Precompute optimal hyperparameters for each method
}
\keyword{internal}
