% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cor.R
\name{estimate_cor}
\alias{estimate_cor}
\title{Compute correlation for fine-mapping}
\usage{
estimate_cor(wmat, ldmat, intercept = F)
}
\arguments{
\item{wmat}{numeric, matrix of weights}

\item{ldmat}{numeric, matrix of linkage disequilibrium}

\item{intercept}{logical, T/F if including an intercept}
}
\value{
correlation between GReX with/without intercept
}
\description{
The function runs a modified version of FOCUS's correlation matrix
estimation
}
