% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap_regress.R
\name{finemap_regress}
\alias{finemap_regress}
\title{Fine-mapping and clustering SEs with SUR}
\usage{
finemap_regress(
  X,
  Y,
  Y.rep,
  R,
  id,
  verbose = F,
  tx_names = NULL,
  coverage = 0.9
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G mean isoform expression across columns}

\item{Y.rep}{matrix, matrix of G isoform expression with replicates}

\item{R}{int, number of replicates}

\item{id}{vector, vector of sample ids showing rep to id}

\item{verbose}{logical}

\item{tx_names}{vector, character vector of tx names in order of columns of Y}
}
\value{
data frame of elastic net, lasso, and LMM based predictions
}
\description{
The function feature selects with SuSiE and runs SUR with cluster-robust
SEs
}
