% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_best_isotwas}
\alias{get_best_isotwas}
\title{Select best model per transcript from multiple isotwas_model objects}
\usage{
get_best_isotwas(model_list, G)
}
\arguments{
\item{model_list}{List of isotwas_model objects}

\item{G}{Number of transcripts}
}
\value{
An isotwas_model with the best model per transcript
}
\description{
Select best model per transcript from multiple isotwas_model objects
}
