% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{get_prediction_matrix}
\alias{get_prediction_matrix}
\title{Get predicted values as a matrix}
\usage{
get_prediction_matrix(model)
}
\arguments{
\item{model}{An isotwas_model object}
}
\value{
A matrix with samples as rows and transcripts as columns
}
\description{
Get predicted values as a matrix
}
