% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resid.R
\name{get_resid}
\alias{get_resid}
\title{Compute residuals for Z-scores}
\usage{
get_resid(zscores, swld, wcor)
}
\arguments{
\item{zscores}{numeric, vector of Z-scores}

\item{swld}{numeric, LD matrix}

\item{wcor}{numeric, correlation matrix from estimate_cor}
}
\value{
residuals and residualized Z-scores
}
\description{
The function runs a modified version of FOCUS's residual calculation
}
