% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{get_weight_matrix}
\alias{get_weight_matrix}
\title{Get weights for all transcripts as a matrix}
\usage{
get_weight_matrix(model, snps = NULL)
}
\arguments{
\item{model}{An isotwas_model object}

\item{snps}{Optional character vector of SNP names to include (default: all)}
}
\value{
A matrix with SNPs as rows and transcripts as columns
}
\description{
Get weights for all transcripts as a matrix
}
