% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_elasticnet.R
\name{multivariate_elasticnet}
\alias{multivariate_elasticnet}
\title{Multivariate elastic net on all columns with row-wise penalty}
\usage{
multivariate_elasticnet(
  X,
  Y,
  Omega,
  scale = F,
  alpha = 0.5,
  nfolds = 5,
  verbose = T,
  par = F,
  n.cores = NULL,
  tx_names = NULL,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{Omega}{matrix, precision matrix of Y}

\item{scale}{logical, T/F to scale Y by Omega}

\item{alpha}{numeric, elastic net mixing parameter}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{seed}{int, random seed}

\item{tx_name}{vector, character vector of tx names in order of columns of Y}
}
\value{
data frame of elastic net, lasso, and LMM based predictions
}
\description{
The function trains multivariate elastic net models for all
isoform transcripts jointly
}
