% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_mmbr.R
\name{multivariate_mmbr}
\alias{multivariate_mmbr}
\title{Multivariate SuSiE model}
\usage{
multivariate_mmbr(
  X,
  Y,
  nfolds = 5,
  verbose = T,
  tx_names = NULL,
  L = 10,
  par,
  n.cores,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical}

\item{tx_names}{vector, character vector of tx names in order of columns of Y}

\item{L}{int, number of single effects}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{seed}{int, random seed}
}
\value{
data frame of mvSuSiE weights
}
\description{
The function trains multivariate SuSiE for all
isoform transcripts jointly
}
