% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_mrmash.R
\name{multivariate_mrmash}
\alias{multivariate_mrmash}
\title{Multivariate MrMash model}
\usage{
multivariate_mrmash(
  X,
  Y,
  nfolds = 5,
  verbose = T,
  tx_names = NULL,
  par,
  n.cores,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{tx_name}{vector, character vector of tx names in order of columns of Y}
}
\value{
mrmash models
}
\description{
The function trains multivariate SuSiE for all
isoform transcripts jointly
}
