% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_spls.R
\name{multivariate_spls}
\alias{multivariate_spls}
\title{Multivariate sparse partial least squares}
\usage{
multivariate_spls(
  X,
  Y,
  nfolds = 5,
  verbose = T,
  tx_names = NULL,
  par,
  n.cores,
  seed
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{tx_name}{vector, character vector of tx names in order of columns of Y}
}
\value{
spls models
}
\description{
The function trains multivariate spls for all
isoform transcripts jointly
}
