% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_confirm.R
\name{p_confirm}
\alias{p_confirm}
\title{Compute confirmation P-value for one gene}
\usage{
p_confirm(p, alpha = 0.05)
}
\arguments{
\item{p}{numeric, vector of TWAS P-values for all isoforms of one gene}

\item{alpha}{numeric, significance level}
}
\value{
confirmation adjusted P-values for all isoforms of one gene
}
\description{
The function runs a modified version of Shaffer's MSRB procedure
to control family-wise error rate for a given gene
}
