% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_r_squared}
\alias{pred_r_squared}
\title{Use anova() to get the sum of squares for the linear model
Calculate the total sum of squares}
\usage{
pred_r_squared(linear.model)
}
\description{
Use anova() to get the sum of squares for the linear model
Calculate the total sum of squares
}
