% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{print.isotwas_model}
\alias{print.isotwas_model}
\title{Print method for isotwas_model}
\usage{
\method{print}{isotwas_model}(x, ...)
}
\arguments{
\item{x}{An isotwas_model object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for isotwas_model
}
