% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{print.isotwas_result}
\alias{print.isotwas_result}
\title{Print method for isotwas_result}
\usage{
\method{print}{isotwas_result}(x, ...)
}
\arguments{
\item{x}{An isotwas_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for isotwas_result
}
