% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{[.isotwas_model}
\alias{[.isotwas_model}
\title{Subset isotwas_model by transcript name}
\usage{
\method{[}{isotwas_model}(x, i)
}
\arguments{
\item{x}{An isotwas_model object}

\item{i}{Transcript name(s) or index}
}
\description{
Subset isotwas_model by transcript name
}
