% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotwas_model.R
\name{summary.isotwas_result}
\alias{summary.isotwas_result}
\title{Summary method for isotwas_result}
\usage{
\method{summary}{isotwas_result}(object, ...)
}
\arguments{
\item{object}{An isotwas_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for isotwas_result
}
