% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas_finemap.R
\name{twas_finemap}
\alias{twas_finemap}
\title{FOCUS-like gene-level finemapping}
\usage{
twas_finemap(res.df, z = "Z", Omega, V, max_enum = 3, cutoff = 0.9)
}
\arguments{
\item{res.df}{data.frame, TWAS results that at least has Z scores and transcript names}

\item{z}{character, column name for Z scores}

\item{Omega}{matrix, matrix of SNP to transcript effect in same order as res.df}

\item{V}{matrix, LD matrix in the same order as Omega}

\item{max_enum}{numeric, max number of causals in the credible set}

\item{cutoff}{numeric, proportion of posterior explained by credible set}
}
\value{
data frame of res.df with pips and whether isoform is in credible set
}
\description{
The function runs probabilistic finemapping on the gene/isoform-level
}
