% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burdenTest.R
\name{burdenTest}
\alias{burdenTest}
\title{Compute weighted burden test}
\usage{
burdenTest(
  mod,
  ld,
  gene,
  sumStats,
  chr,
  pos,
  a1,
  a2,
  a1_mod = "ALT",
  a2_mod = "REF",
  snpName = "SNP",
  Z = NULL,
  beta = NULL,
  se = NULL,
  featureName = "Feature",
  R2cutoff = 0.01,
  alpha = 2.5e-06,
  nperms = 1000,
  usePos = F
)
}
\arguments{
\item{mod}{data.frame, model for a given isoform}

\item{ld}{matrix, ld reference matrix}

\item{gene}{character, gene name}

\item{sumStats}{data frame, GWAS summary statistics}

\item{chr}{character, colnames in sumStats that keeps the chromosome}

\item{pos}{character, colnames in sumStats that keeps the position}

\item{a1}{character, colnames in sumStats that keeps the ALT allele}

\item{a2}{character, colnames in sumStats that keeps the REF allele}

\item{a1_mod}{character, colnames in model that keeps the ALT allele}

\item{a2_mod}{character, colnames in model that keeps the REF allele}

\item{snpName}{character, colnames in sumStats that keeps the SNP id}

\item{Z}{character, colnames in sumStats that keeps the Z score}

\item{beta}{character, colnames in sumStats that keeps the effect size}

\item{se}{character, colnames in sumStats that keeps the standard error}

\item{featureName}{character, colname in model that keeps the feature name}

\item{R2cutoff}{numeric, predictive R2 cutoff}

\item{alpha}{numeric, P-value threshold for permutation testing}

\item{nperms}{numeric, number of permutations}

\item{usePos}{logical, use SNP positions vs. SNP ids}
}
\value{
list of results for burden and permutation tests
}
\description{
The function takes in a gene expression model in MOSTWAS form
and GWAS summary statistics and
carries out the weighted burden Z-test for a trait
}
