% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isotwas.R
\name{compute_isotwas}
\alias{compute_isotwas}
\title{Compute isoTWAS model for a set of isoforms/transcripts}
\usage{
compute_isotwas(
  X,
  Y,
  Y.rep,
  R,
  id,
  omega_est = "replicates",
  omega_nlambda = 10,
  method = c("mrce_lasso", "curds_whey", "multi_enet", "finemap", "univariate",
    "mvsusie"),
  predict_nlambda = 50,
  family = "gaussian",
  scale = F,
  alpha = 0.5,
  nfolds = 5,
  verbose = F,
  par = F,
  n.cores = NULL,
  tx_names = NULL,
  seed = NULL,
  run_all = T,
  return_all = F,
  tol.in = 0.001,
  maxit.in = 1000,
  coverage = 0.9
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{Y.rep}{matrix, matrix of G isoform expression with replicates}

\item{R}{int, number of replicates}

\item{id}{vector, vector of sample ids showing rep to id}

\item{omega_est}{character, 'replicates' or 'mean' to use Y.rep or Y}

\item{omega_nlambda}{int, number of omegas to generate}

\item{method}{character, vector of methods to use}

\item{predict_nlambda}{int, number of lambdas in MRCE}

\item{family}{character, glmnet family}

\item{scale}{logical, T/F to scale Y by Omega}

\item{alpha}{numeric, elastic net mixing parameter}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{tx_names}{vector, character vector of tx names - order of columns of Y}

\item{seed}{int, random seed}

\item{return_all}{logical, return R2 for all models?}

\item{tol.in}{numeric, tolerance for objective difference}

\item{maxit.in}{int, maximum number of iteractions}
}
\value{
optimal isoTWAS model
}
\description{
The function runs MVR models for a set of transcripts and outputs
the best model
}
