% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isotwas.R
\name{compute_isotwas}
\alias{compute_isotwas}
\title{Compute isoTWAS model for a set of isoforms/transcripts}
\usage{
compute_isotwas(
  X,
  Y,
  Y.rep,
  R,
  id,
  Omega = NULL,
  omega_est = "replicates",
  omega_nlambda = 15,
  method = c("mrce_lasso", "curds_whey", "multi_enet", "mrce_enet", "finemap",
    "uni_enet", "uni_susie", "uni_blup"),
  family = "gaussian",
  scale = F,
  alpha = 0.5,
  nfolds = 5,
  verbose,
  par = F,
  n.cores = NULL,
  tx_names = NULL,
  seed = NULL
)
}
\arguments{
\item{X}{matrix, design matrix of SNP dosages}

\item{Y}{matrix, matrix of G isoform expression across columns}

\item{Y.rep}{matrix, matrix of G isoform expression with replicates}

\item{R}{int, number of replicates}

\item{id}{vector, vector of sample ids showing rep to id}

\item{family}{character, glmnet glm family}

\item{scale}{logical, T/F to scale Y by Omega}

\item{alpha}{numeric, elastic net mixing parameter}

\item{nfolds}{int, number of CV folds}

\item{verbose}{logical}

\item{par}{logical, uses mclapply to parallelize model fit}

\item{n.cores}{int, number of parallel cores}

\item{tx_names}{vector, character vector of tx names in order of columns of Y}
}
\value{
optimal isoTWAS model
}
\description{
The function runs MVR models for a set of transcripts and outputs
the best model
}
